feat(types): Add Note Type and Resolver, Add Place Resolvers
Added a new Note type definition and a resolver to retrieve them from a Photo. Flickr strangely does not have a getList method for notes, they're instead added to the data returned by photo.getInfo instead. This makes for a roundabout resolver that would benefit from some sort of request caching strategy. Otherwise two calls will be made to photo.getInfo if the notes field needs to be resolved on Photo.

Added resolvers for the Place type and hooked them up to location field in the Photo type.