fix(dotenv): Add .env Setup Script
Added a script to create a .env file from the template and added it as an install task in .travis.yml. This was needed to eliminate errors from inline-dotenv, which could not find a .evn file present during CI builds.

Reorganized .env.default into Required and Optional settings, sorted alphabetically