fix(coverage): Move Babel 'test' config back to .babelrc
Having this as part of AVA's config in package.json was causing coverage to not be generated, as nyc needed the babel settings as well.

Removed node versions 7 and 6 from .travis.yml because catbox is not compatible with those versions of node. This may explain why caching with catbox did not work, because babel was transpiling down to node 6.10.